Functionality changes
Changed the way sound is played through the notification:
A popup window will now be opened and remain open unless the user closes it. This window will be reused to play sound, and will automatically be created if it has been previously closed. At most one of these windows will exist at any given time, and can be minimized.

Removed point totals from goal descriptors because if the API call is made shortly after the goal was scored in real time, then it may be initially inaccurate.